requirements.txt
Flask~=1.1.2
Pillow~=8.0.1
pytz~=2020.5
requests~=2.23.0
python-dotenv~=0.17.1
fake-headers
